[Silver IV] Title: 비트가 넘쳐흘러, Time: 80 ms, Memory: 11480 KB -BaekjoonHub

+++ 백준/Silver/17419. 비트가 넘쳐흘러/비트가 넘쳐흘러.java
@@ -6,7 +6,7 @@ public class Main{
 	public static void main(String[] args) throws IOException{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int k = Integer.parseInt(br.readLine());
-		int result = Integer.parseInt(br.readLine(),2);
+		long result = Long.parseLong(br.readLine(),2);
 		int cnt = 0;
 		while (result>0) {
 			cnt++;
